Fix several functions in queue code
Add the implementation of three more basic functions in queue code.

As "make check" cannot successfully insert, remove, and clean the queue.
There are also memory referencing errors when the program try to clean the queue.
Put the "entry" pointer inside the loop, to access the entry memory, and free them using
the given API call.

Due to "entry" pointer being outside of the loop, it was never referenced.